remove event-specific import
