--- test/buildWorkshop/importMainWorkshop.test.ts
@@ -1,7 +1,0 @@
-import { importMainWorkshop } from '../../src/buildWorkshop/importMainWorkshop';
-
-describe('importProducts', () => {
-  test('Main Workshop Wiki Vers', () => {
-    expect(() => importMainWorkshop(true)).not.toThrowError();
-  });
-});
